get rehype/remark working
