+++ app/lib/projects.ts
@@ -1,5 +1,4 @@
 import fs from 'fs/promises'
-import matter from 'gray-matter'
 import path from 'path'
 
 export async function getAllProjects(): Promise<any[]> {
@@ -9,51 +8,9 @@ export async function getAllProjects(): Promise<any[]> {
   const projects = await Promise.all(
     mdxFiles.map(async (fileName) => {
       const slug = fileName.replace(/\.mdx$/, '')
-      const source = await fs.readFile(path.join(dir, fileName), 'utf8')
-      const { data } = matter(source)
-      return {
-        slug,
-        name: data.name as string,
-        date: data.date as string,
-        summary: data.summary as string | undefined,
-        image: {
-          src: data.image?.src as string | undefined,
-          alt: data.image?.alt as string | undefined,
-          width: data.image?.width as number | undefined,
-          height: data.image?.height as number | undefined,
-        },
-        description: data.description as string | undefined,
-        tags: data.tags as string[] | undefined,
-      }
+      const { metadata } = await import(`@/projects/${slug}.mdx`)
+      return { slug, ...metadata }
     })
   )
   return projects.sort((a, b) => new Date(b.date).getTime() - new Date(a.date).getTime())
 }
-
-export async function getProjectBySlug(slug: string): Promise<any> {
-  const dir = path.join(process.cwd(), 'app/projects')
-  const filePath = path.join(dir, `${slug}.mdx`)
-
-  try {
-    const source = await fs.readFile(filePath, 'utf8')
-    const { data } = matter(source)
-    return {
-      slug,
-      name: data.name as string,
-      date: data.date as string,
-      summary: data.summary as string | undefined,
-      image: {
-        src: data.image?.src as string | undefined,
-        alt: data.image?.alt as string | undefined,
-        width: data.image?.width as number | undefined,
-        height: data.image?.height as number | undefined,
-      },
-      description: data.description as string | undefined,
-      tags: data.tags as string[] | undefined,
-    }
-  } catch (error) {
-    console.error(`Error reading project file for slug "${slug}":`, error)
-    return null
-  }
-}
-
