+++ app/projects/homelab.mdx
@@ -1,13 +1,14 @@
----
-name:         Homelab
-description:  Ongoing project of managing servers and services at home.
-image:
-  src:        /img/projects/homelab/massimo-botturi-zFYUsLk_50Y-unsplash.jpg
-  alt:        Homelab
-  width:      2268
-  height:     4032
-date:         "2020-02-25T00:00:00Z"
----
+export const metadata = {
+  name: 'Homelab',
+  description: 'Ongoing project of managing servers and services at home.',
+  image: {
+    src: '/img/projects/homelab/massimo-botturi-zFYUsLk_50Y-unsplash.jpg',
+    alt: 'Homelab',
+    width: 2268,
+    height: 4032,
+  },
+  date: '2020-02-25T00:00:00Z',
+}
 
 An ongoing project for services I run at home. I have a few Intel NUCs running [Proxmox](https://www.proxmox.com/), a few Synology NAS for persistent storage and backups, and a handful of Raspberry Pis for ad-hoc projects.
