+++ app/projects/iheart-hackweek-2019.mdx
@@ -1,13 +1,14 @@
----
-name:         iHeart Hackweek 2019
-description:  Spent a week in a New York mansion working on a side project with some coworkers.
-image:
-  src:        /img/projects/iheart-hackweek-2019/iHeartRadio_Logo_iHR_Vertical_Color.png
-  alt:        iHeartRadio Logo
-  width:      1680
-  height:     1305
-date:         "2019-07-31T00:00:00Z"
----
+export const metadata = {
+  name: 'iHeart Hackweek 2019',
+  description: 'Spent a week in a New York mansion working on a side project with some coworkers.',
+  image: {
+    src: '/img/projects/iheart-hackweek-2019/iHeartRadio_Logo_iHR_Vertical_Color.png',
+    alt: 'iHeartRadio Logo',
+    width: 1680,
+    height: 1305,
+  },
+  date: '2019-07-31T00:00:00Z',
+}
 
 I participated in the iHeartRadio (my company's sister company) hackweek in July 2019. While I can't go in-depth on the team's project, it was great spending time with coworkers in a different environment and learning about other teams' projects.
