--- app/projects/launchpad.md
@@ -1,16 +1,0 @@
----
-name:         Launchpad
-description:  A side project I did at work to showcase innovation at the company as part of a new corporate website launch.
-image:
-  src:        /img/projects/launchpad/launchpad.png
-  alt:        Map of radio stations
-  width:      3584
-  height:     2240
-date:         "2018-05-18T00:00:00Z"
-tags:
-  - data visualization
----
-
-A side project I did at work to showcase innovation at the company as part of a new corporate website launch.
-
-Project available at [Jelli Launchpad](https://www.jelli.com/launchpad/). Click on the map.
