+++ app/projects/loudness-wars.mdx
@@ -1,17 +1,15 @@
----
-name:         Loudness Wars
-description:  A closer look at how recorded music has been getting increasingly louder over the past 50 years.
-image:
-  src:        /img/projects/loudness-wars/mick-haupt-CbNBjnXXhNg-unsplash.jpg
-  alt:        A stack of records at a record store
-  width:      5472
-  height:     3648
-date:         "2021-07-05T00:00:00Z"
-tags:
-  - data visualization
-  - python
-  - music
----
+export const metadata = {
+  name: 'Loudness Wars',
+  description: 'A closer look at how recorded music has been getting increasingly louder over the past 50 years.',
+  image: {
+    src: '/img/projects/loudness-wars/mick-haupt-CbNBjnXXhNg-unsplash.jpg',
+    alt: 'A stack of records at a record store',
+    width: 5472,
+    height: 3648,
+  },
+  date: '2021-07-05T00:00:00Z',
+  tags: ['data visualization', 'python', 'music'],
+}
 
 The [Loudness Wars](https://en.wikipedia.org/wiki/Loudness_war) refers to the dramatic increase in the perceived loudness of a song over the past 40 years. The loud parts haven't necessarily become louder, but instead the effective loudness (called the root mean square, or RMS) of recorded music has increased. This has been done by compressing the loud parts of an audio signal, and then increasing the overall loudness of the signal. The loud parts are now more or less as loud as they were before dynamic compression, but now that overall loudness has increased.
