+++ app/projects/paperless-ng-terraform.mdx
@@ -1,16 +1,15 @@
----
-name:         paperless-ng Terraform
-description:  Terraform module to set up paperless-ng on AWS.
-image:
-  src:        /img/projects/paperless-ng-terraform/wesley-tingey-snNHKZ-mGfE-unsplash.jpg
-  alt:        A stack of books, binders, and folders
-  width:      6000
-  height:     4000
-date:         "2023-01-22T00:00:00Z"
-tags:
-  - aws
-  - terraform
----
+export const metadata = {
+  name: 'paperless-ng Terraform',
+  description: 'Terraform module to set up paperless-ng on AWS.',
+  image: {
+    src: '/img/projects/paperless-ng-terraform/wesley-tingey-snNHKZ-mGfE-unsplash.jpg',
+    alt: 'A stack of books, binders, and folders',
+    width: 6000,
+    height: 4000,
+  },
+  date: '2023-01-22T00:00:00Z',
+  tags: ['aws', 'terraform'],
+}
 
 This module creates the necessary resources to run [paperless-ng](https://paperless-ng.readthedocs.io/en/latest/index.html) on AWS.
