+++ app/projects/terraform-provider-glinet.mdx
@@ -1,16 +1,15 @@
----
-name:         GL.iNet Terraform Provider
-description:  A Terraform provider for interacting with GL.iNet routers
-image:
-  src:        /img/projects/terraform-provider-glinet/axt1800_scene_cabin.jpg
-  alt:        GL.iNet AXT1800 Travel Router
-  width:      1920
-  height:     1080
-date:         2023-11-26
-tags:
-  - terraform
-  - homelab
----
+export const metadata = {
+  name: 'GL.iNet Terraform Provider',
+  description: 'A Terraform provider for interacting with GL.iNet routers',
+  image: {
+    src: '/img/projects/terraform-provider-glinet/axt1800_scene_cabin.jpg',
+    alt: 'GL.iNet AXT1800 Travel Router',
+    width: 1920,
+    height: 1080,
+  },
+  date: '2023-11-26',
+  tags: ['terraform', 'homelab'],
+}
 
 I have a nifty little [GL.iNet travel router](https://www.gl-inet.com/products/gl-axt1800/) that I use when traveling. To ensure a secure connection, I've automated the process of connecting it to my VPN in my [homelab](/projects/homelab). This not only maintains a secure connection but also allows me to leverage services I run at home, such as a network-wide ad blocker.
