+++ app/projects/treefort-2019.mdx
@@ -1,13 +1,14 @@
----
-name:         Hackfort 2019 - Creative Outlets Inspire Innovation
-description:  I gave a talk on side projects and innovation at a music-meets-tech festival in Boise.
-image:
-  src:        /img/projects/treefort-2019/hackfort.png
-  alt:        Hackfort logo
-  width:      225
-  height:     225
-date:         "2019-03-27T00:00:00Z"
----
+export const metadata = {
+  name: 'Hackfort 2019 - Creative Outlets Inspire Innovation',
+  description: 'I gave a talk on side projects and innovation at a music-meets-tech festival in Boise.',
+  image: {
+    src: '/img/projects/treefort-2019/hackfort.png',
+    alt: 'Hackfort logo',
+    width: 225,
+    height: 225,
+  },
+  date: '2019-03-27T00:00:00Z',
+}
 
 In March 2019, I presented at [Hackfort](https://www.treefortmusicfest.com/fort/hackfort/), part of the Treefort music festival in Boise, Idaho. My panel was on side projects and innovation, and how side projects foster a mindset of innovation that lends itself nicely to day-to-day work.
