--- next.config.js
@@ -1,28 +1,0 @@
-/* eslint-env node */
-const createMDX = require('@next/mdx')
-
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-  pageExtensions: ['md', 'mdx', 'ts', 'tsx'],
-  async redirects() {
-    return [
-      {
-        source: '/:year((?!_next).*)/:month/:day/:post.html',
-        destination: '/blog/:year-:month-:day-:post',
-        permanent: false,
-      },
-    ]
-  },
-  transpilePackages: [
-    'd3',
-
-    // transitive dependencies of d3
-    'delaunator',
-    'internmap',
-    'robust-predicates',
-  ],
-}
-
-const withMdx = createMDX()
-
-module.exports = withMdx(nextConfig)
